Latest SRC 22.09

+++ Tera Emulator v1xxx VS Project [SRC]/TERA_C++_EMULATOR_1/Entity.h
@@ -6,5 +6,5 @@ class Entity
 public :
 	Entity();
 	virtual ~Entity();
-	unsigned long long _entityId;
+	unsigned int _entityId;
 };
